requirements.txt
fastapi
uvicorn
pandas
spotipy
python-dotenv
openai
httpx
itsdangerous